// src/app/crypto-currency/model/crypto-currency-transactions.model.ts
export class CryptoCurrencyTransactions {
    cryptocurrencytransactionId: number;
    senderFrom: string;
    currencyFrom: string;
    destinationTo: string;
    currencyTo: string;
    amount: number;
    storageLocation: string;
    notice: string;
    attachment: boolean;
    attachmentPath: string;
    attachmentName: string;
    attachmentType: string;
}